initial commit for game api
create game api shell
