+++ assets/scripts/auth/ui.js
@@ -27,26 +27,26 @@ const signInFailure = function (error) {
 }
 
 const changePasswordSuccess = function () {
-  $('#message').text('Change password successfully')
-  $('#message').css('background-color', 'green')
+  $('#messageTwo').text('Change password successfully')
+  $('#messageTwo').css('background-color', 'green')
   console.log('change password successfully')
 }
 
 const changePasswordFailure = function (error) {
-  $('#message').text('Error on changing password')
-  $('#message').css('background-color', 'red')
+  $('#messageTwo').text('Error on changing password')
+  $('#messageTwo').css('background-color', 'red')
   console.log(error)
 }
 
 const signOutSuccess = function () {
-  $('#message').text('Signed out successfully')
-  $('#message').css('background-color', 'green')
+  $('#messageTwo').text('Signed out successfully')
+  $('#messageTwo').css('background-color', 'green')
   console.log('signed out successfully')
 }
 
 const signOutFailure = function (error) {
-  $('#message').text('Error on signing out')
-  $('#message').css('background-color', 'red')
+  $('#messageTwo').text('Error on signing out')
+  $('#messageTwo').css('background-color', 'red')
   console.log(error)
 }
