+++ assets/scripts/index.js
@@ -3,20 +3,22 @@
 const setAPIOrigin = require('../../lib/set-api-origin')
 const config = require('./config')
 const authEvents = require('./auth/events.js')
-const example = require('./example')
+// const example = require('./example')
 
 $(() => {
   setAPIOrigin(location, config)
 })
 
+const gameLogic = require('./gameLogic.js')
+
 $(() => {
   // player click -> function 'switchPlayer' happens
   // create a new Games on the server
   authEvents.addHandlers()
-  example.gameEvent()
-  example.reset()
-  example.signIn()
-  example.signOut()
+  gameLogic.gameEvent()
+  gameLogic.reset()
+  gameLogic.signIn()
+  gameLogic.signOut()
 })
 
 // use require with a reference to bundle the file and use it in this file
