+++ index.html
@@ -13,6 +13,7 @@
     <body class="container-fluid">
 
     <!--sign up page-->
+        <div id="first-page">
         <h1 class="title">Tic Tac Toe Game</h1>
 
           <p id="message"></p>
@@ -33,12 +34,12 @@
             <button id="in-button" type="submit" value="Sign in">Sign in!</button>
         </fieldset>
         </form>
+      </div>
 
         <div id="gamePage" class="gamePage">
           <h1 class="title">Tic Tac Toe Game!</h1>
           <p class="intro">Tic-tac-toe is a simple & FUN game for two players, X and O <br> who take turns marking the spaces in a 3×3 grid.<br> The player who succeeds in placing three of their marks in a horizontal, vertical, or diagonal row wins the game. </p>
-        <!--<div id=game-board> -->
-        <!--Game Page-->
+         <p id="messageTwo"></p>
           <form id="change-password" class="game-auth">
           <fieldset>
             <legend>Change Password?</legend>
@@ -67,7 +68,7 @@
           </div>
           <form class="message">next turn is ...</form>
 
-          <div>
+          <div class="gameBoard">
             <ul class="gameBoard">
              <li id="0" class="box one">#</li>
              <li id="1" class="box two">#</li>
