+++ scripts/games/create.sh
@@ -6,18 +6,18 @@ curl "https://aqueous-atoll-85096.herokuapp.com/games" \
   --include \
   --request POST \
   --header "Content-Type: application/json" \
-  --header "Authorization: Token token=<token>" \
+  --header "Authorization: Token token=${TOKEN}>" \
   --data '{
-  "game": {
-    "id": 3,
-    "cells": ["","","","","","","","",""],
-    "over": false,
-    "player_x": {
-      "id": 1,
-      "email": "k@gmail.com"
-    },
-    "player_o": null
-  }
-}'
+    "game": {
+      "id": "",
+      "cells": ["","","","","","","","",""],
+      "over": false,
+      "player_x": {
+        "id": ""
+        "email": ""
+      },
+      "player_o": null
+    }
+  }'
 
 echo
